refactor(ui): distribute tests.rs inline and remove central test file
